update permissions, copywriting, vars
